Restore backend database folder renamings - those changes got undone somehow
