+++ src/backend/db/index.js
@@ -5,7 +5,7 @@ import configuration from './configuration';
 import users from './users';
 import featurePermissions from './feature-permissions';
 
-const connection = require('../../../knexfile')[process.env.NODE_ENV || 'test'];
+const connection = require('../knexfile')[process.env.NODE_ENV || 'test'];
 // eslint-disable-next-line global-require
 export const db = require('knex')(connection);
